+++ service/controller/v10/resource/s3bucket/resource.go
@@ -14,7 +14,7 @@ import (
 const (
 	// Name is the identifier of the resource.
 	Name = "s3bucketv10"
-	// Lifecycle ID for the logging bucket
+	// LifecycleLoggingBucketID is the Lifecycle ID for the logging bucket
 	LifecycleLoggingBucketID = "ExpirationLogs"
 )
 
@@ -27,7 +27,7 @@ type Config struct {
 
 	// Settings.
 	InstallationName     string
-	AccessLogsExpiration int64
+	AccessLogsExpiration int
 }
 
 // DefaultConfig provides a default configuration to create a new s3bucket
@@ -50,7 +50,7 @@ type Resource struct {
 
 	// Settings.
 	installationName     string
-	accessLogsExpiration int64
+	accessLogsExpiration int
 }
 
 // New creates a new configured s3bucket resource.
